Fix bug with long files and reduce memory leak
The long file bug was due to type of FileLenght, int is to short and loops.
Change it into long long unsigned int to be sure.

I find out there was huge memory leak due to the wrong eliminination of
SSegment->ptrArray. TestDeleteSegment is a destructor function that take
care of SSegment.